+++ BITS/Get all BITS jobs, custom.ps1
@@ -1,36 +1,29 @@
 ﻿#Requires -RunAsAdministrator
 
-$OutputTemplate1 = @'
-{0}
-{1}
-{2} | {3} | {4}
-{5:N0} / {6:N0}
-Files:
-'@
+# TO DO: Rewrite using list view
 
-$OutputTemplate2 =@'
+function Main {
+  $OutputTemplate1 = "{0}`n{1}`n{2} | {3} | {4}`n{5:N0} / {6:N0}`nFiles:"
+  $OutputTemplate2 = "    Remote Name: {0}`n    Local Name:  {1}`n    Progress:    {2:N0} / {3:N0} ({4})"
+  $SplitterLine = [String]::new([char]0x2500, 110)
 
-    Remote Name: {0}
-    Local Name:  {1}
-    Progress:    {2:N0} / {3:N0} ({4})
-'@
+  # The following only works in PowerShell 7.x:
+  # $SplitterLine = [String]::new("`u{2500}", 110)
 
-$SplitterLine = [String]::new([char]0x2500, 110)
-
-# The following only works in PowerShell 7.x:
-# $SplitterLine = [String]::new("`u{2500}", 110)
-
-function Public_Static_Void_Main {
+  $a = Get-BitsTransfer -AllUsers -ErrorAction Stop
+  if ($null -eq $a) {
+    Write-Output "Found no BITS jobs."
+    break
+  }
   Write-Output $SplitterLine
-  Get-BitsTransfer -AllUsers -ErrorAction Stop | ForEach-Object {
-      $a=$_
-      Write-Output $($OutputTemplate1 -f $a.JobId,$a.DisplayName,$a.TransferType,$a.JobState,$a.Priority,$a.BytesTransferred,$a.BytesTotal)
-      ForEach-Object -InputObject $a.FileList {
-          if ($_.IsTransferComplete) { $b = "Complete" } else { $b = "Pending" }
-          Write-Output $($OutputTemplate2 -f $_.RemoteName,$_.LocalName,$_.BytesTransferred,$_.BytesTotal,$b)
-      }
-      Write-Output $SplitterLine
+  $a | ForEach-Object {
+    Write-Output $($OutputTemplate1 -f $_.JobId, $_.DisplayName, $_.TransferType, $_.JobState, $_.Priority, $_.BytesTransferred, $_.BytesTotal)
+    ForEach-Object -InputObject $_.FileList {
+      if ($_.IsTransferComplete) { $b = "Complete" } else { $b = "Pending" }
+      Write-Output $($OutputTemplate2 -f $_.RemoteName, $_.LocalName, $_.BytesTransferred, $_.BytesTotal, $b)
+    }
+    Write-Output $SplitterLine
   }
 }
 
-Public_Static_Void_Main
+Main